Changed ExpectedPayloadByutes to be uint64. This is actually wrong but it needs to be 8 bytes or it will mismatch with the messaging system. Added TODO to resolve this somewhere else later

+++ TubesTypes.h
@@ -47,8 +47,8 @@ struct Listener {
 struct ReceiveBuffer {
 	ReceiveBuffer();
 
-	int16_t ExpectedHeaderBytes;	// How many more bytes of header we expect for the current packet // TODODB: 
-	int32_t ExpectedPayloadBytes;	// How many more bytes of payload we expect for the current packet
+	int16_t ExpectedHeaderBytes;	// How many more bytes of header we expect for the current packet
+	uint64_t ExpectedPayloadBytes;	// How many more bytes of payload we expect for the current packet  // TODODB: This variable should be int32_t since the recv can't take more in a single call
 	Byte*	PayloadData;			// Dynamic buffer for packet payload
 	Byte*	Walker;					// Pointer to where the next recv should write to
 };
